requirements.txt
google-genai
pinecone-client
psycopg2
python-dotenv
langchain-text-splitters
numpy
pandas
